requirements.txt
absl-py>=0.7.1
astor>=0.8.0
astroid>=2.2.5
autopep8>=1.4.4
beautifulsoup4>=4.8.0
bs4>=0.0.1
certifi>=2019.6.16
chardet>=3.0.4
cloudpickle>=1.2.1
colorama>=0.4.1
cycler>=0.10.0
cytoolz>=0.10.0
dask>=2.2.0
decorator>=4.4.0
Django>=2.2.1
django-cors-headers>=3.0.2
django-filter>=2.2.0
django-guardian>=2.0.0
django-otp>=0.6.0
djangorestframework>=3.10.1
gast>=0.2.2
grpcio>=1.16.1
h5py>=2.9.0
idna>=2.8
imageio>=2.5.0
isort>=4.3.20
kiwisolver>=1.1.0
lazy-object-proxy>=1.4.1
Markdown>=3.1.1
matplotlib>=3.1.0
mccabe>=0.6.1
mkl-fft>=1.0.14
mkl-random>=1.0.2
mkl-service>=2.0.2
networkx>=2.3
numpy>=1.16.4
olefile>=0.46
opencv-contrib-python>=4.1.0.25
opencv-python>=4.1.0.25
Pillow>=6.0.0
protobuf>=3.8.0
pycodestyle>=2.5.0
Pygments>=2.4.2
PyJWT>=1.7.1
pylint>=2.3.1
pyparsing>=2.4.0
pyreadline>=2.1
PySocks>=1.7.0
python-dateutil>=2.8.0
pytz>=2019.1
PyWavelets>=1.0.3
PyYAML>=5.1.1
requests>=2.22.0
rest-condition>=1.0.3
scikit-image>=0.15.0
scipy>=1.3.0
sip>=4.19.13
six>=1.12.0
soupsieve>=1.9.2
sqlparse>=0.3.0
termcolor>=1.1.0
tesserocr>=2.3.1
toolz>=0.10.0
tornado>=6.0.3
twilio>=6.29.2
typed-ast>=1.4.0
urllib3>=1.25.3
Werkzeug>=0.15.4
wincertstore>=0.2
wrapt>=1.11.2
